Created a responsive layout for movie detail page and cart page. Works pretty well for most screen widths. Also added local storage for the userCart.
